private_warehouse_django/requirements.txt
certifi==2015.9.6.2
clarify-python==2.0.0
decorator==4.0.4
Django==1.8.4
django-extensions==1.5.7
django-taggit==0.17.1
djangorestframework==3.2.4
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
Pillow==2.9.0
psycopg2==2.6.1
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
urllib3==1.12
wheel==0.24.0
